iqtisodiy_yordamchi/requirements.txt
Flask==2.3.0
Flask-SQLAlchemy==3.0.5
Flask-Babel==3.1.0
Werkzeug==2.3.0
SQLAlchemy==2.0.0
openpyxl==3.1.5
reportlab==4.0.7
twilio==8.10.0
jsonschema==4.18.0
statsmodels==0.14.0
scipy==1.11.0
openai==1.3.0
celery==5.3.0
redis==5.0.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9
requests==2.31.0
flask-socketio==5.3.2
python-telegram-bot==20.0
pytesseract==0.3.10
Pillow==10.0.0

# Mobile & Desktop Support
i18next==0.2.0
axios==0.27.2

